Create a separate AsterixConn for AsterixDB-related connections
